some more minor bug fixes and beautification
git-svn-id: https://svn.atnf.csiro.au/difx@273 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ utilities/difx2fits/trunk/src/fitsAG.c
@@ -83,7 +83,7 @@ const DifxInput *DifxInput2FitsAG(const DifxInput *D,
 	if(nRowBytes != sizeof(row))
 	{
 		fprintf(stderr, "AG table : nRowBytes != sizeof(row) : "
-			"%d != %u\n", nRowBytes, sizeof(row));
+			"%d != %u\n", nRowBytes, (unsigned int)(sizeof(row)));
 		exit(0);
 	}
